# Make thread statics faster