This is a thing I noticed when I profiled the TechEmpower plaintext benchmark and it's a piece of pretty low hanging fruit.

* Fix the TODO to inline `.cctor` check into the R2R helper. This avoids a some method calls and stack frame setup.
* Remove array cast from the hot path. The runtime doesn't care what type the storage is, as long as it's a managed reference - in fact, it seems to be treated as an `Object` in `pCurrentThread->GetThreadStaticStorageForModule` anyway. I considered replacing the cast with an `Unsafe.As`, but just redeclaring the signatures as returning `object[]` seemed better (because now we actually guarantee the classlib won't pass anything but `object[]` to `RhSetThreadStaticStorageForModule`).

For a simple program:

```csharp
class Program
{
    [ThreadStatic]
    static int X; // = 456;

    [MethodImpl(MethodImplOptions.NoInlining)]
    static void Set() => X = 123;

    static void Main(string[] args)
    {
        var sw = Stopwatch.StartNew();

        for (int i = 0; i < 10000000; i++)
        {
            Set();
        }

        Console.WriteLine(sw.ElapsedMilliseconds);
    }
}
```

This leads to following improvements:

|     | Before | After |
| ------------- |:-------------:| -----:|
| Program.cs without static constructor | 70 ms | 45 ms |
| Program.cs with static constructor | 164 ms | 46 ms |

The running times are comparable with the desktop CLR now.