@mattwarren 2018-07-18
@MichalStrehovsky BTW if you find yourself writing more benchmarks, you might want to take a look at [BenchmarkDotNet](https://benchmarkdotnet.org/), especially as @adamsitnik recently added [CoreRT support](https://github.com/dotnet/BenchmarkDotNet/issues/718) (allows CoreRT v. CoreCLR comparisons) and it has lots of [other really nice features](https://benchmarkdotnet.org/Overview.htm)